requirements/base.txt
# Wheel 0.25+ needed to install certain packages on CPython 3.5+
# like Pillow and psycopg2
wheel==0.29.0

django==1.10.4

# Configuration
django-environ==0.4.0

# Python-PostgreSQL Database Adapter
psycopg2==2.6.2

# Images
Pillow==3.4.2

djangorestframework==3.5.3
djangorestframework-gis==0.11
